refactor: move d4 to vectors
