+++ src/coordinax/__init__.py
@@ -11,7 +11,6 @@ from .setup_package import RUNTIME_TYPECHECKER
 with install_import_hook("coordinax", RUNTIME_TYPECHECKER):
     from . import angle, distance, frames, operators
     from ._src import (
-        d4,
         dn,
         exceptions,
         funcs,
@@ -20,7 +19,6 @@ with install_import_hook("coordinax", RUNTIME_TYPECHECKER):
         utils,
         vectors,
     )
-    from ._src.d4 import *
     from ._src.dn import *
     from ._src.exceptions import *
     from ._src.funcs import *
@@ -31,6 +29,7 @@ with install_import_hook("coordinax", RUNTIME_TYPECHECKER):
     from ._src.vectors.d1 import *
     from ._src.vectors.d2 import *
     from ._src.vectors.d3 import *
+    from ._src.vectors.d4 import *
     from ._version import version as __version__
     from .distance import Distance
 
@@ -49,7 +48,7 @@ __all__ += vectors.base.__all__
 __all__ += vectors.d1.__all__
 __all__ += vectors.d2.__all__
 __all__ += vectors.d3.__all__
-__all__ += d4.__all__
+__all__ += vectors.d4.__all__
 __all__ += dn.__all__
 __all__ += space.__all__
 __all__ += exceptions.__all__
@@ -65,7 +64,6 @@ del (
     transform,
     typing,
     utils,
-    d4,
     dn,
     funcs,
     RUNTIME_TYPECHECKER,
