--- src/coordinax/_src/d4/__init__.py
@@ -1,12 +1,0 @@
-"""4-dimensional."""
-# pylint: disable=duplicate-code
-
-from . import base, compat, spacetime
-from .base import *
-from .compat import *
-from .spacetime import *
-
-__all__: list[str] = []
-__all__ += base.__all__
-__all__ += spacetime.__all__
-__all__ += compat.__all__
