+++ src/coordinax/_src/operators/galilean/boost.py
@@ -15,10 +15,10 @@ import quaxed.numpy as jnp
 import unxt as u
 
 from .base import AbstractGalileanOperator
-from coordinax._src.d4.spacetime import FourVector
 from coordinax._src.operators.base import AbstractOperator, op_call_dispatch
 from coordinax._src.operators.identity import IdentityOperator
 from coordinax._src.vectors.d3 import AbstractPos3D, CartesianVel3D
+from coordinax._src.vectors.d4 import FourVector
 
 
 @final
