+++ src/coordinax/_src/vectors/d4/__init__.py
@@ -1,0 +1,14 @@
+"""4-dimensional."""
+
+__all__ = [
+    # Base
+    "AbstractPos4D",
+    # Spacetime
+    "FourVector",
+]
+
+from .base import AbstractPos4D
+from .spacetime import FourVector
+
+# isort: split
+from . import compat  # noqa: F401
