+++ src/coordinax/_src/vectors/d4/compat.py
