+++ src/coordinax/_src/vectors/d4/spacetime.py
